Fixes dependency on Brave DRM.
Fixes brave/brave-browser#5749

We make blink/renderer/core depend on Brave DRM, however, core is a
`component` which, in component build, becomes a shared library. This
means Brave DRM needs to be exported from the library. To achieve that
we mark the code generated from mojom with export class attribute for
blink core and use code generated for blink. The code that consumes
Brave DRM should use the blink generated header and namespace.